revision1

+++ SSFR01/foodapp/src/main/java/tutorial/seho/foodapp/util/RedisUtil.java
@@ -2,8 +2,11 @@ package tutorial.seho.foodapp.util;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.math.BigDecimal;
+
 
 import tutorial.seho.foodapp.model.Food;
+
 
 public class RedisUtil {
 
@@ -12,7 +15,7 @@ public class RedisUtil {
         List<Food> foodList = new LinkedList<>();
         for(String item : dbQuery){
             String[] itemStr = item.split("-");
-            foodList.add(new Food(itemStr[0],Double.parseDouble(itemStr[1])));
+            foodList.add(new Food(itemStr[0],BigDecimal.valueOf(Double.parseDouble(itemStr[1]))));
         }
         return foodList;
     }
@@ -25,7 +28,7 @@ public class RedisUtil {
 
     public static void addEgFoodnoDB(List<Food> foodList, int qty){
         for(int i = 0; i < qty; i++){
-            Food newFood = new Food("Dish"+Integer.toString(i), (double)(i*i));
+            Food newFood = new Food("Dish"+Integer.toString(i), BigDecimal.valueOf((double)(i*i)));
             foodList.add(newFood);
         }  
     }
